+++ core/src/elements/ons-switch.js
@@ -66,8 +66,8 @@ export default class SwitchElement extends BaseCheckboxElement {
       this.attributeChangedCallback('modifier', null, this.getAttribute('modifier'));
     });
 
-    this._boundOnChange = this._onChange.bind(this);
-    this._boundOnRelease = this._onRelease.bind(this);
+    this._onChange = this._onChange.bind(this);
+    this._onRelease = this._onRelease.bind(this);
   }
 
   get _scheme() {
@@ -117,15 +117,18 @@ export default class SwitchElement extends BaseCheckboxElement {
   }
 
   _onClick(ev) {
-    if (ev.target.classList.contains(`${this.defaultElementClass}__touch`)) {
+    if (ev.target.classList.contains(`${this.defaultElementClass}__touch`)
+      || ev.timeStamp === this._lastTimeStamp // Prevent second click triggered by <label>
+    ) {
       ev.preventDefault();
     }
+    this._lastTimeStamp = ev.timeStamp;
   }
 
   _onHold(e) {
     if (!this.disabled) {
       ModifierUtil.addModifier(this, 'active');
-      document.addEventListener('release', this._boundOnRelease);
+      document.addEventListener('release', this._onRelease);
     }
   }
 
@@ -141,7 +144,7 @@ export default class SwitchElement extends BaseCheckboxElement {
     this._startX = this._locations[this.checked ? 1 : 0];// - e.gesture.deltaX;
 
     this.addEventListener('drag', this._onDrag);
-    document.addEventListener('release', this._boundOnRelease);
+    document.addEventListener('release', this._onRelease);
   }
 
   _onDrag(e) {
@@ -162,7 +165,7 @@ export default class SwitchElement extends BaseCheckboxElement {
     }
 
     this.removeEventListener('drag', this._onDrag);
-    document.removeEventListener('release', this._boundOnRelease);
+    document.removeEventListener('release', this._onRelease);
 
     this._handle.style.left = '';
     ModifierUtil.removeModifier(this, 'active');
@@ -185,7 +188,7 @@ export default class SwitchElement extends BaseCheckboxElement {
 
   connectedCallback() {
     contentReady(this, () => {
-      this._input.addEventListener('change', this._boundOnChange);
+      this._input.addEventListener('change', this._onChange);
     });
 
     this.addEventListener('dragstart', this._onDragStart);
@@ -197,7 +200,7 @@ export default class SwitchElement extends BaseCheckboxElement {
 
   disconnectedCallback() {
     contentReady(this, () => {
-      this._input.removeEventListener('change', this._boundOnChange);
+      this._input.removeEventListener('change', this._onChange);
     });
 
     this.removeEventListener('dragstart', this._onDragStart);