+++ CHANGELOG.md
@@ -12,6 +12,7 @@ dev
  * ons-navigator: Fixed[#2286](https://github.com/OnsenUI/OnsenUI/issues/2286).
  * ons-navigator: Fixed[#1992](https://github.com/OnsenUI/OnsenUI/issues/1992).
  * css-components: Fixed [#2045](https://github.com/OnsenUI/OnsenUI/issues/2045).
+ * ons-switch: Prevents two click events in a row when placed inside a 'label' tag.
 
  ### Misc